Document and group subclasses
