+++ src/main/scala/pokemon/model/Trainer.scala
@@ -10,6 +10,9 @@ abstract class Trainer {
 
   def chooseMove(): Move
 
+  /**
+    * Generates a deck of 3 random Pokemon
+    */
   def generateDeck(): Unit = {
     val allPokemons = PokemonRegistry.pokemons
     val rand = new Random()
@@ -21,6 +24,11 @@ abstract class Trainer {
     }
   }
 
+  /**
+    * Hard limit of 3 Pokemon per Trainer
+    *
+    * @param pokemon
+    */
   def addPokemon(pokemon: Pokemon): Unit = {
     if (this.deck.size >= 3) throw new Exception(s"$name can only have 3 Pokemon")
     this.deck += pokemon
@@ -42,4 +50,31 @@ abstract class Trainer {
   }
 
   def isDefeated: Boolean = deck.forall(_.currentHP == 0)
+}
+
+class Player extends Trainer {
+  val name: String = "Player"
+  private var _moveIndex: Int = -1
+
+  def moveIndex(index: Int): Unit = this._moveIndex = index
+
+  override def chooseMove(): Move = activePokemon.moves(this._moveIndex)
+}
+
+class Bot extends Trainer {
+  val name: String = "Bot"
+
+  override def chooseMove(): Move = {
+    randomMove()
+  }
+
+  /**
+    * Randomly selects a move from the active Pokemon's move list
+    *
+    * @return
+    */
+  private def randomMove(): Move = {
+    val moveIndex = Random.nextInt(activePokemon.moves.length)
+    activePokemon.moves(moveIndex)
+  }
 }
